# src/main/resources/META-INF/openapi.yaml
...
openapi: 3.0.3
info:
  title: Points-of-Interest Gateway API
  description: API surface exposed to interact with Backends (summit POIs + parks data) via the Gateway
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://book02-poi-map-app-book02.apps.cluster-8l688.sandbox1215.opentlc.com/
    description: (backstage component deployment)
  - url: http://localhost:8080/
    description: (localhost testing)
paths:
...
  /backend/register/{backendId}:
    get:
      tags:
      - Backend Resource
      parameters:
      - name: backendId
        in: path
        required: true
        schema:
          type: string
      - name: endpoint
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
  /backend/unregister/{backendId}:
    get:
      tags:
      - Backend Resource
      parameters:
      - name: backendId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...